Don't override the main repo if it's already there
For testing purposes one might want to use the nightly channel with the
oneclick installer. Right now it consists of changing the pahkat config
and then running the oneclick installer but it was consistently
resetting the channel back to stable, giving mixed results and hard to
debug discrepancies. This makes it so that if the main channel is
already present then it won't touch any repo configuration.